+++ src/main/java/com/whiskels/notifier/telegram/service/DemoUserService.java
@@ -19,7 +19,7 @@ public class DemoUserService extends UserService {
     }
 
     @Override
-    public User getOrCreate(int chatId) {
+    public User getOrCreate(Long chatId) {
         return get(chatId)
                 .orElseGet(() -> {
                     User user = new User(chatId);
